% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_zcss_deps.R
\name{add_zcss_deps}
\alias{add_zcss_deps}
\title{Add zuericssstyle CSS dependencies to an HTML tag}
\usage{
add_zcss_deps(tag)
}
\arguments{
\item{tag}{A htmltools tag or tagList (for example the UI returned by
\code{shiny::fluidPage()}). The returned value is a \code{tagList} that includes
the provided \code{tag} and the package dependencies.}
}
\value{
A \code{tagList} containing \code{tag} plus the CSS htmlDependencies.
}
\description{
Attach the package CSS (from \verb{inst/css/}) as html dependencies to a tag or tagList. This
ensures styles are available when rendering HTML,
pkgdown sites, htmlwidgets, or Shiny apps.
}
\examples{
\dontrun{
inst / examples / ssz_deps_with_bootstrap / app.r
}

}
