% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_css.R
\name{get_css_core}
\alias{get_css_core}
\title{get_css_core}
\usage{
get_css_core(origin, destination_filepath)
}
\arguments{
\item{origin}{path to internal css, "css/sszThemeShiny.css" or "css/sszTheme.css"}

\item{destination_filepath}{where to copy the file to}
}
\description{
checks if destination filepath is valid and copies chosen css file to destination
}
