% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_css.R
\name{get_generic_css}
\alias{get_generic_css}
\title{get_generic_css}
\usage{
get_generic_css(destination_filepath = "ssz.css")
}
\arguments{
\item{destination_filepath}{optional, defaults to "ssz.css" in the local
folder. if provided, give full path and filename, like "path/folder/file.css"}
}
\description{
function to create a css file for the generic styling of html/shiny according to the CI/CD of the city of Zurich
}
\examples{
get_generic_css()
}
