% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_css.R
\name{get_shiny_css}
\alias{get_shiny_css}
\title{get_shiny_css}
\usage{
get_shiny_css(destination_filepath = "sszThemeShiny.css")
}
\arguments{
\item{destination_filepath}{optional, defaults to "sszThemeShiny.css" in the local
folder. if provided, give full path and filename, like "path/folder/file.css"}
}
\description{
creates a css file for html and shiny widget styling
}
\examples{
get_shiny_css()
}
