% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssz_page.R
\name{ssz_page_without_bootstrap}
\alias{ssz_page_without_bootstrap}
\title{ssz_page_without_bootstrap}
\usage{
ssz_page_without_bootstrap(...)
}
\arguments{
\item{...}{ui content}
}
\value{
ui with ssz styling applied
}
\description{
Function to use instead of fluidPage that has the ssz css styling applied and the bootstrap css removed
}
