% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sszActionButton.R
\name{sszActionButton}
\alias{sszActionButton}
\title{sszActionButton}
\usage{
sszActionButton(inputId, label, ...)
}
\arguments{
\item{inputId}{inputId parameter for actionButton}

\item{label}{label parameter for actionButton}

\item{...}{further parameters for actionButton}
}
\value{
the return of actionButton with "sszActionButton" added as a class
}
\description{
function to generate an action button with the custom ssz styling
}
\examples{
library(shiny)

ui <- add_zcss_deps(
  fluidPage(
    h1("Widget demo"),
    sszActionButton(
      "action",
      "Abfrage starten",
    ),
    br(),
    h1("Action"),
    textOutput("action")
  )
)

server <- function(input, output, session) {
  # Observe the action button press
  observeEvent(input$action, {
    output$action <- renderText({
      "Abfrage gestartet"
    })
  })
}

shinyApp(ui, server)
}
