% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssz_ui_utils.R
\name{sszActionButtonIcon}
\alias{sszActionButtonIcon}
\title{sszActionButtonIcon}
\usage{
sszActionButtonIcon(inputId, label, ssz_icon, ...)
}
\arguments{
\item{inputId}{inputId parameter for actionButton}

\item{label}{label parameter for actionButton}

\item{ssz_icon}{htmltools::tags$img(...) with an icon/image to be used on the button}

\item{...}{further parameters for actionButton}
}
\value{
the return of actionButton with "sszActionButton" added as a class
}
\description{
sszActionButtonIcon
}
\examples{
\dontrun
{sszActionButtonIcon("ActionButtonId",
"Abfrage starten",
ssz_icon = img(icons_ssz$`arrow-right`)
)}
}
