% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sszAutocompleteInput.R
\name{sszAutocompleteInput}
\alias{sszAutocompleteInput}
\title{sszAutocompleteInput}
\usage{
sszAutocompleteInput(inputId, label, options, ...)
}
\arguments{
\item{inputId}{inputId parameter for autocomplete_input}

\item{label}{label parameter for autocomplete_input}

\item{options}{options parameter for autocomplete_input}

\item{...}{further parameters for autocomplete_input}
}
\value{
same as autocomplete_input
}
\description{
sszAutocompleteInput
}
\examples{
\dontrun{
inst / examples / sszAutocompleteInput / app.r
}
}
\seealso{
Other shiny_widgets: 
\code{\link{sszActionButton}()},
\code{\link{sszAirDatepickerInput}()},
\code{\link{sszDateRange}()},
\code{\link{sszDownloadButton}()},
\code{\link{sszNumericInput}()},
\code{\link{sszOgdDownload}()},
\code{\link{sszRadioButtons}()},
\code{\link{sszRadioGroupButtons}()},
\code{\link{sszSelectInput}()},
\code{\link{sszSliderInput}()},
\code{\link{sszTextInput}()}
}
\concept{shiny_widgets}
