% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sszDateRange.R
\name{sszDateRange}
\alias{sszDateRange}
\title{sszDateRange}
\usage{
sszDateRange(inputId, label, format = "dd.mm.yyyy", language = "de", ...)
}
\arguments{
\item{inputId}{inputId to dateRangeInput}

\item{label}{label parameter for dateRangeInput}

\item{format}{optional format parameter for dateRangeInput, default "dd.mm.yyyy"}

\item{language}{optional language parameter for dateRangeInput, default "de"}

\item{...}{further named parameters for dateRangeInput (e.g. start, end, min, max, separator)}
}
\value{
the same as dateRangeInput return, with modified label class
}
\description{
function to generate a DateRangeInput with the custom ssz label styling (via "ssz-label" class)
}
\examples{
\dontrun{
inst / examples / sszDateRange / app.r
}
}
\seealso{
Other shiny_widgets: 
\code{\link{sszActionButton}()},
\code{\link{sszAirDatepickerInput}()},
\code{\link{sszAutocompleteInput}()},
\code{\link{sszDownloadButton}()},
\code{\link{sszNumericInput}()},
\code{\link{sszOgdDownload}()},
\code{\link{sszRadioButtons}()},
\code{\link{sszRadioGroupButtons}()},
\code{\link{sszSelectInput}()},
\code{\link{sszSliderInput}()},
\code{\link{sszTextInput}()}
}
\concept{shiny_widgets}
