% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssz_ui_utils.R
\name{sszDownload}
\alias{sszDownload}
\title{sszDownload}
\usage{
sszDownload(outputId, label = "Download")
}
\arguments{
\item{outputId}{shiny outputID for the download button}

\item{label}{shiny label for the download button}
}
\value{
tag a object
}
\description{
function to generate a ssz-themed download button for csv and excel
}
\examples{
sszDownload("csvDownload", label = "csv")
}
