% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sszDownloadButton.R
\name{sszDownloadButton}
\alias{sszDownloadButton}
\title{sszDownloadButton}
\usage{
sszDownloadButton(outputId, label, image = NULL, icon = NULL, ...)
}
\arguments{
\item{outputId}{shiny outputID for the download button}

\item{label}{shiny label for the download button}

\item{image}{html-i tag with an image/icon to be added before the text; default NULL}

\item{icon}{optional icon parameter for downloadButton, default NULL}

\item{...}{optional further named params for downloadButton (class parameter will be ignored/overwritten)}
}
\value{
a downloadButton with class sszDownload and optionally the image added
}
\description{
function to generate a ssz-themed download button for csv and excel, wraps around downloadButton
}
\examples{
library(shiny)

icons_ssz <- icons::icon_set(here::here("inst", "examples", "icons"))

ui <- fluidPage(
  includeCSS(here::here("inst", "css", "sszThemeShiny.css")),
  h1("Widget demo"),
  tags$div(
    id = "downloadWrapperId",
    class = "downloadWrapperDiv",
    sszDownloadButton("csvDownload",
                      label = "csv",
                      image = img(icons_ssz("download"))
    )
  )
)

server <- function(input, output, session) {
  # Download data
  data <- mtcars

  output$csvDownload <- downloadHandler(
    filename = function() {
      paste("data-", Sys.Date(), ".csv", sep = " ")
    },
    content = function(file) {
      write.csv(data, file)
    }
  )

}

shinyApp(ui, server)
}
