% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sszOgdDownload.R
\name{sszOgdDownload}
\alias{sszOgdDownload}
\title{sszOgdDownload}
\usage{
sszOgdDownload(
  outputId,
  href,
  label = "OGD",
  image = icons_ssz("external-link")
)
}
\arguments{
\item{outputId}{outputId parameter}

\item{href}{link to be opened}

\item{label}{label parameter (i.e. text to be shown), default is "OGD"}

\item{image}{optional image/icon in the form of a html-i Tag, default is icons_ssz("external-link")}
}
\value{
tags$a
}
\description{
function to generate an ssz-themed ogd download button
}
\examples{
\dontrun{
inst / examples / sszOgdDownload / app.r
}
}
\seealso{
Other shiny_widgets: 
\code{\link{sszActionButton}()},
\code{\link{sszAirDatepickerInput}()},
\code{\link{sszAutocompleteInput}()},
\code{\link{sszDateRange}()},
\code{\link{sszDownloadButton}()},
\code{\link{sszNumericInput}()},
\code{\link{sszRadioButtons}()},
\code{\link{sszRadioGroupButtons}()},
\code{\link{sszSelectInput}()},
\code{\link{sszSliderInput}()},
\code{\link{sszTextInput}()}
}
\concept{shiny_widgets}
