% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sszRadioButtons.R
\name{sszRadioButtons}
\alias{sszRadioButtons}
\title{sszRadioButtons}
\usage{
sszRadioButtons(inputId, label, choices, ...)
}
\arguments{
\item{inputId}{inputId parameter for radioButtons}

\item{label}{label parameter for radioButtons}

\item{choices}{choices parameter for radioButtons}

\item{...}{further named parameters for radioBuTons (e.g. inline = TRUE for horizontal layout)}
}
\value{
the return of radioButtons with modified labels, wrapped in a div with class "radioDiv"
}
\description{
function to generate radioButtons with the custom ssz (label) styling
}
\details{
a "sszRadioButton" is added to the output of radioButtons, and a
"sszRadioButton-label" class is added to the appropriate child. the whole thing is
wrapped in a div with class "radioDiv"
}
\examples{
library(shiny)
ui <- fluidPage(
  includeCSS(here::here("inst", "css", "sszThemeShiny.css")),
  h1("Widget demo"),
  sszRadioButtons(
    inputId = "ButtonGroupLabel",
    label = "Flughafen",
    choices = c("HOU", "LAX", "JFK"),
    selected = "JFK" # default value
  ),
  br(),
  h1("Selected"),
  textOutput("choice")
)

server <- function(input, output, session) {
  # Our dataset
  output$choice <- renderText(input$ButtonGroupLabel)
}
shinyApp(ui, server)
}
