% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sszRadioGroupButtons.R
\name{sszRadioGroupButtons}
\alias{sszRadioGroupButtons}
\title{sszRadioGroupButtons}
\usage{
sszRadioGroupButtons(
  inputId,
  label = NULL,
  choices = NULL,
  selected = NULL,
  ...
)
}
\arguments{
\item{inputId}{inputId parameter for radioButtons}

\item{label}{label parameter for radioButtons, default NULL}

\item{choices}{choices parameter for radioButtons, default NULL}

\item{selected}{initially selected value, default NULL}

\item{...}{further named parameters for shinyWidget::radioGroupButtons}
}
\value{
the return of radioGroupButtons with modified labels
}
\description{
function to generate radioButtons with the custom ssz (label) styling
}
\details{
a "sszRadioButton" is added to the output of radioButtons, and a
"sszRadioButton-label" class is added to the appropriate child. the whole thing is
wrapped in a div with class "radioDiv"
}
\examples{
\dontrun{
inst / examples / sszRadioGroupButtons / app.r
}
}
