% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssz_ui_utils.R
\name{sszSelectInput}
\alias{sszSelectInput}
\title{sszSelectInput}
\usage{
sszSelectInput(inputId, label, choices, ...)
}
\arguments{
\item{inputId}{inputId parameter for selectInput}

\item{label}{label parameter for selectInput}

\item{choices}{choices parameter for selectInput,}

\item{...}{further named parameters to be passed to selectInput, e.g. selected}
}
\value{
the same as selectInput with modified label class
}
\description{
function to generate a selectInput with the custom ssz label styling (via "ssz-label" class)
}
\examples{
sszSelectInput("select", "Destination:", choices = unique(data$dest), selected = "LAX")
}
