% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sszSliderInput.R
\name{sszSliderInput}
\alias{sszSliderInput}
\title{sszSliderInput}
\usage{
sszSliderInput(inputId, label, value, min, max, ...)
}
\arguments{
\item{inputId}{inputId parameter for sliderInput}

\item{label}{label parameter for sliderInput}

\item{value}{value parameter (initial value) for sliderInput}

\item{min}{min value that can be selected for sliderInput}

\item{max}{max value that can be selected for sliderInput}

\item{...}{further named parameters to be passed to sliderInput (step/ticks/width)}
}
\value{
the same as textInput with modified label class
}
\description{
function to generate a sliderInput with the custom ssz label styling (via "ssz-label" class)
}
\examples{
library(shiny)

ui <- fluidPage(
  includeCSS(here::here("inst", "css", "sszThemeShiny.css")),
  h1("Widget demo"),
  sszSliderInput("choose_number",
                 "Input:",
                 value = c(30, 60),
                 min = 0,
                 max = 100),
  br(),
  h1("Input"),
  textOutput("input")
)

server <- function(input, output, session) {
  # Our dataset
  output$input <- renderText(input$choose_number)
}

shinyApp(ui, server)
}
