% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sszTextInput.R
\name{sszTextInput}
\alias{sszTextInput}
\title{sszTextInput}
\usage{
sszTextInput(inputId, label, ...)
}
\arguments{
\item{inputId}{inputId parameter for textInput}

\item{label}{label parameter for textInput}

\item{...}{further named parameters passed to textInput}
}
\value{
the same as textInput with modified label class
}
\description{
function to generate a textInput with the custom ssz label styling (via "ssz-label" class)
}
\examples{
library(shiny)

ui <- fluidPage(
  includeCSS(here::here("inst", "css", "sszThemeShiny.css")),
  h1("Widget demo"),
  sszTextInput("suchfeld",
               "Name:"),
  br(),
  h1("Input"),
  textOutput("input")
)

server <- function(input, output, session) {
  # Our dataset
  output$input <- renderText(input$suchfeld)
}
shinyApp(ui, server)
}
