% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_zcss_deps.R
\name{add_zcss_deps}
\alias{add_zcss_deps}
\title{add_zcss_deps}
\usage{
add_zcss_deps(tag)
}
\arguments{
\item{tag}{tag/tagList, e.g. a fluidPage}
}
\value{
fluidPage/tag with ssz css dependencies added, including shiny widget stylings
}
\description{
add_zcss_deps
}
\examples{
\dontrun{
inst / examples / ssz_deps_with_bootstrap / app.r
}

}
