% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssz_page.R
\name{ssz_page}
\alias{ssz_page}
\title{ssz_page}
\usage{
ssz_page(...)
}
\arguments{
\item{...}{ui content}
}
\value{
ui with ssz styling applied
}
\description{
Function to use instead of fluidPage that has the ssz css styling applied
}
\examples{
\dontrun{inst/examples/sszActionButton/app.r}
}
