% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sszActionButton.R
\name{sszActionButton}
\alias{sszActionButton}
\title{sszActionButton}
\usage{
sszActionButton(inputId, label, ...)
}
\arguments{
\item{inputId}{inputId parameter for actionButton}

\item{label}{label parameter for actionButton}

\item{...}{further parameters for actionButton}
}
\value{
the return of actionButton with "sszActionButton" added as a class
}
\description{
function to generate an action button with the custom ssz styling
}
\examples{
\dontrun{
inst / examples / sszActionButton / app.r
}
}
\seealso{
Other shiny_widgets: 
\code{\link{sszAirDatepickerInput}()},
\code{\link{sszAutocompleteInput}()},
\code{\link{sszDateRange}()},
\code{\link{sszDownloadButton}()},
\code{\link{sszNumericInput}()},
\code{\link{sszOgdDownload}()},
\code{\link{sszRadioButtons}()},
\code{\link{sszRadioGroupButtons}()},
\code{\link{sszSelectInput}()},
\code{\link{sszSliderInput}()},
\code{\link{sszTextInput}()}
}
\concept{shiny_widgets}
