% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sszActionButtonIcon.R
\name{sszActionButtonIcon}
\alias{sszActionButtonIcon}
\title{sszActionButtonIcon}
\usage{
sszActionButtonIcon(inputId, label, ssz_icon, ...)
}
\arguments{
\item{inputId}{inputId parameter for actionButton}

\item{label}{label parameter for actionButton}

\item{ssz_icon}{htmltools::tags$img(...) with an icon/image to be used on the button}

\item{...}{further parameters for actionButton}
}
\value{
the return of actionButton with "sszActionButton" added as a class
}
\description{
function to generate an action button with an icon and with the custom ssz styling
}
\examples{
\dontrun{inst/examples/sszActionButtonIcon/app.r}
}
