% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sszActionButtonIcon.R
\name{sszActionButtonIcon}
\alias{sszActionButtonIcon}
\title{sszActionButtonIcon}
\usage{
sszActionButtonIcon(inputId, label, ssz_icon, ...)
}
\arguments{
\item{inputId}{inputId parameter for actionButton}

\item{label}{label parameter for actionButton}

\item{ssz_icon}{htmltools::tags$img(...) with an icon/image to be used on the button}

\item{...}{further parameters for actionButton}
}
\value{
the return of actionButton with "sszActionButton" added as a class
}
\description{
function to generate an action button with an icon and with the custom ssz styling
}
\examples{
library(shiny)

icons_ssz <- icons::icon_set(here::here("inst", "examples", "icons"))

ui <- fluidPage(
  includeCSS(here::here("inst", "css", "sszThemeShiny.css")),
  h1("Widget demo"),
  sszActionButtonIcon(
    "action",
    "Abfrage starten",
    ssz_icon = img(icons_ssz("arrow-right"))
  ),
  br(),
  h1("Action"),
  textOutput("action")
)

server <- function(input, output, session) {
  # Observe the action button press
  observeEvent(input$action, {
    output$action <- renderText({
      "Abfrage gestartet"
    })
  })
}

shinyApp(ui, server)
}
