% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssz_ui_utils.R
\name{sszAutocompleteInput}
\alias{sszAutocompleteInput}
\title{sszAutocompleteInput}
\usage{
sszAutocompleteInput(inputId, label, options, ...)
}
\arguments{
\item{inputId}{inputId parameter for autocomplete_input}

\item{label}{label parameter for autocomplete_input}

\item{options}{options parameter for autocomplete_input}

\item{...}{further parameters for autocomplete_input}
}
\value{
same as autocomplete_input
}
\description{
sszAutocompleteInput
}
\examples{
\dontrun{
sszAutocompleteInput(
 "street",
 "Geben Sie eine Strasse ein",
 unique(addresses$StrasseLang)
)
}
}
