% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sszAutocompleteInput.R
\name{sszAutocompleteInput}
\alias{sszAutocompleteInput}
\title{sszAutocompleteInput}
\usage{
sszAutocompleteInput(inputId, label, options, ...)
}
\arguments{
\item{inputId}{inputId parameter for autocomplete_input}

\item{label}{label parameter for autocomplete_input}

\item{options}{options parameter for autocomplete_input}

\item{...}{further parameters for autocomplete_input}
}
\value{
same as autocomplete_input
}
\description{
sszAutocompleteInput
}
\examples{
library(shiny)
opts <- sapply(1:100, function(i) paste0(sample(letters, 7), collapse=""))
ui <- fluidPage(
  includeCSS(here::here("inst", "css", "sszThemeShiny.css")),
  h1("Widget demo"),
  sszAutocompleteInput(
    "select",
    "Geben Sie einen Namen ein",
    opts
  ),
  br(),
  h1("Selected"),
  textOutput("choice")
)

server <- function(input, output, session) {
  # Our dataset
  output$choice <- renderText(input$select)
}
shinyApp(ui, server)
}
