% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sszDateRange.R
\name{sszDateRange}
\alias{sszDateRange}
\title{sszDateRange}
\usage{
sszDateRange(inputId, label, format = "dd.mm.yyyy", language = "de", ...)
}
\arguments{
\item{inputId}{inputId to dateRangeInput}

\item{label}{label parameter for dateRangeInput}

\item{format}{optional format parameter for dateRangeInput, default "dd.mm.yyyy"}

\item{language}{optional language parameter for dateRangeInput, default "de"}

\item{...}{further named parameters for dateRangeInput (e.g. start, end, min, max, separator)}
}
\value{
the same as dateRangeInput return, with modified label class
}
\description{
function to generate a DateRangeInput with the custom ssz label styling (via "ssz-label" class)
}
\examples{
library(shiny)
library(icons)
library(here)

icons_ssz <- icon_set(here("man/icons"))

ui <- fluidPage(
  includeCSS(here::here("inst", "css", "sszThemeShiny.css")),
  h1("Widget demo"),
  sszDateRange("DateRange", "Datum:",
               start  = "2001-01-01",
               end    = "2010-12-31",
               min    = "2001-01-01",
               max    = "2012-12-21",
               separator = icons_ssz("calendar")
  ),
  br(),
  h1("Selected"),
  textOutput("choice")
)

server <- function(input, output, session) {
  # Our dataset
  output$choice <- renderText(format(input$DateRange, "\%d.\%m.\%Y"))
}
shinyApp(ui, server)
}
