% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sszDownloadButton.R
\name{sszDownloadButton}
\alias{sszDownloadButton}
\title{sszDownloadButton}
\usage{
sszDownloadButton(outputId, label, image = NULL, icon = NULL, ...)
}
\arguments{
\item{outputId}{shiny outputID for the download button}

\item{label}{shiny label for the download button}

\item{image}{html-i tag with an image/icon to be added before the text; default NULL}

\item{icon}{optional icon parameter for downloadButton, default NULL}

\item{...}{optional further named params for downloadButton (class parameter will be ignored/overwritten)}
}
\value{
a downloadButton with class sszDownload and optionally the image added
}
\description{
function to generate a ssz-themed download button for csv and excel, wraps around downloadButton
}
\examples{
\dontrun{inst/examples/sszDownloadButton/app.r}
}
