% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sszNumericInput.R
\name{sszNumericInput}
\alias{sszNumericInput}
\title{sszNumericInput}
\usage{
sszNumericInput(inputId, label, value, ...)
}
\arguments{
\item{inputId}{inputId parameter for numericInput}

\item{label}{label parameter for numericInput}

\item{value}{value parameter (initial value) for numericInput}

\item{...}{further named parameters to be passed to numericInput (min/max/step/width)}
}
\value{
the same as textInput with modified label class
}
\description{
function to generate a numericInput with the custom ssz label styling (via "ssz-label" class)
}
\examples{
\dontrun{inst/examples/sszNumericInput/app.r}
}
