% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssz_ui_utils.R
\name{sszOgdDownload}
\alias{sszOgdDownload}
\title{sszOgdDownload}
\usage{
sszOgdDownload(inputId, label, onclick)
}
\arguments{
\item{inputId}{inputId parameter for actionButton}

\item{label}{label parameter for actionButton}

\item{onclick}{onclick parameter for actionButton}
}
\value{
same as the shiny actionButton but with the "sszDownload" added as a class
}
\description{
function to generate an ssz-themed ogd download button
}
\examples{
sszOgdDownload(inputId = "ogdDown", label = "OGD",
onclick = "window.open('https://data.stadt-zuerich.ch/', '_blank')")
}
