% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sszOgdDownload.R
\name{sszOgdDownload}
\alias{sszOgdDownload}
\title{sszOgdDownload}
\usage{
sszOgdDownload(outputId, href, label, image = NULL)
}
\arguments{
\item{outputId}{outputId parameter}

\item{href}{link to be opened}

\item{label}{label parameter (i.e. text to be shown)}

\item{image}{optional image/icon in the form of a html-i Tag, default NULL}
}
\value{
tags$a
}
\description{
function to generate an ssz-themed ogd download button
}
\examples{
library(shiny)

icons_ssz <- icons::icon_set(here::here("inst", "examples", "icons"))

ui <- fluidPage(
  includeCSS(here::here("inst", "css", "sszThemeShiny.css")),
  h1("Widget demo"),
  tags$div(
    id = "downloadWrapperId",
    class = "downloadWrapperDiv",
    sszOgdDownload(outputId = "ogdDown",
                   label = "OGD",
                   href = "https://data.stadt-zuerich.ch/",
                   image = img(icons_ssz("external-link"))
    )
  )
)

server <- function(input, output, session) {
}

shinyApp(ui, server)
}
