% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssz_ui_utils.R
\name{sszOgdDownload}
\alias{sszOgdDownload}
\title{sszOgdDownload}
\usage{
sszOgdDownload(outputId, href, label, image = NULL)
}
\arguments{
\item{outputId}{outputId parameter}

\item{href}{link to be opened}

\item{label}{label parameter (i.e. text to be shown)}

\item{image}{optional image/icon in the form of a html-i Tag, default NULL}
}
\value{
tags$a
}
\description{
function to generate an ssz-themed ogd download button
}
\examples{
sszOgdDownload(
  outputId = "ogdDown", label = "OGD",
  href = "https://data.stadt-zuerich.ch"
)
}
