% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sszRadioGroupButtons.R
\name{sszRadioGroupButtons}
\alias{sszRadioGroupButtons}
\title{sszRadioGroupButtons}
\usage{
sszRadioGroupButtons(
  inputId,
  label = NULL,
  choices = NULL,
  selected = NULL,
  ...
)
}
\arguments{
\item{inputId}{inputId parameter for radioGroupButtons}

\item{label}{label parameter for radioGroupButtons, default NULL; leave NULL for button above chart (i.e. no label)}

\item{choices}{choices parameter for radioGroupButtons, default NULL}

\item{selected}{initially selected value, default NULL}

\item{...}{further named parameters for shinyWidget::radioGroupButtons}
}
\value{
the return of radioGroupButtons with modified html classes
}
\description{
function to generate radioGroupButtons from shinyWidgets with the custom ssz (label) styling
}
\details{
classes ssz-radiogroup-1, ssz-radiogroup-2, and ssz-radiogroup-3 are added
(starting from outermost div, moving down 1 step at a time); though currently these are not used for styling;
only .btn-group-toggle>.btn-default and .btn-default.active are used for the styling
}
\examples{
\dontrun{
inst / examples / sszRadioGroupButtons / app.r
}
}
