% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sszSelectInput.R
\name{sszSelectInput}
\alias{sszSelectInput}
\title{sszSelectInput}
\usage{
sszSelectInput(inputId, label, choices, selectize = TRUE, ...)
}
\arguments{
\item{inputId}{inputId parameter for selectInput}

\item{label}{label parameter for selectInput}

\item{choices}{choices parameter for selectInput,}

\item{selectize}{optional selectize parameter, default TRUE, but watch out for influence of other widgets (related to this? https://github.com/rstudio/shiny/issues/3125)}

\item{...}{further named parameters to be passed to selectInput, e.g. selected}
}
\value{
the same as selectInput with modified label class
}
\description{
function to generate a selectInput with the custom ssz label styling (via "ssz-label" class)
}
\examples{
\dontrun{
inst / examples / sszSelectInput / app.R
}
}
\seealso{
Other shiny_widgets: 
\code{\link{sszActionButton}()},
\code{\link{sszAirDatepickerInput}()},
\code{\link{sszAutocompleteInput}()},
\code{\link{sszDateRange}()},
\code{\link{sszDownloadButton}()},
\code{\link{sszNumericInput}()},
\code{\link{sszOgdDownload}()},
\code{\link{sszRadioButtons}()},
\code{\link{sszRadioGroupButtons}()},
\code{\link{sszSliderInput}()},
\code{\link{sszTextInput}()}
}
\concept{shiny_widgets}
