% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssz_ui_utils.R
\name{sszSliderInput}
\alias{sszSliderInput}
\title{sszSliderInput}
\usage{
sszSliderInput(inputId, label, value, min, max, ...)
}
\arguments{
\item{inputId}{inputId parameter for sliderInput}

\item{label}{label parameter for sliderInput}

\item{value}{value parameter (initial value) for sliderInput}

\item{min}{min value that can be selected for sliderInput}

\item{max}{max value that can be selected for sliderInput}

\item{...}{further named parameters to be passed to sliderInput (step/ticks/width)}
}
\value{
the same as textInput with modified label class
}
\description{
function to generate a sliderInput with the custom ssz label styling (via "ssz-label" class)
}
\examples{
sszSliderInput(inputId = "grade_number", label = "Grade:", value = 5, min = 1, max = 6)
}
