% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sszTextInput.R
\name{sszTextInput}
\alias{sszTextInput}
\title{sszTextInput}
\usage{
sszTextInput(inputId, label, ...)
}
\arguments{
\item{inputId}{inputId parameter for textInput}

\item{label}{label parameter for textInput}

\item{...}{further named parameters passed to textInput}
}
\value{
the same as textInput with modified label class
}
\description{
function to generate a textInput with the custom ssz label styling (via "ssz-label" class)
}
\examples{
\dontrun{
inst / examples / sszTextInput / app.r
}
}
\seealso{
Other shiny_widgets: 
\code{\link{sszActionButton}()},
\code{\link{sszAirDatepickerInput}()},
\code{\link{sszAutocompleteInput}()},
\code{\link{sszDateRange}()},
\code{\link{sszDownloadButton}()},
\code{\link{sszNumericInput}()},
\code{\link{sszOgdDownload}()},
\code{\link{sszRadioButtons}()},
\code{\link{sszRadioGroupButtons}()},
\code{\link{sszSelectInput}()},
\code{\link{sszSliderInput}()}
}
\concept{shiny_widgets}
