% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zuericssstyle-package.R
\docType{package}
\name{zuericssstyle-package}
\alias{zuericssstyle}
\alias{zuericssstyle-package}
\title{zuericssstyle: Statistik Stadt Zürich CSS styling for shiny apps and html}
\description{
Style shiny apps according to the city of Zurich's corporate design.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/StatistikStadtZuerich/zuericssstyle}
}

}
\author{
\strong{Maintainer}: Sara Hänzi \email{sara.haenzi@zuerich.ch}

Authors:
\itemize{
  \item Alessandro Feller \email{alessandro.feller@zuerich.ch}
}

Other contributors:
\itemize{
  \item Statistik stadt Zürich \email{statistik@zuerich.ch} [copyright holder, funder]
}

}
\keyword{internal}
